Modify UI

+++ src/components/ModuleForum.tsx
@@ -14,20 +14,28 @@ import {
   deleteSubscription,
 } from "../redux/features/users/userSlice";
 import { WhiteLoader } from "./Loader";
+import { isLoggedIn } from "../utils";
+import CombinedAuthenticationPage from "../pages/CombinedAuthenticationPage";
+import { toggleLogin } from "../redux/features/modal/modal";
 
-export const RedButton = styled.button`
+export const Button = styled.button<{subscribed?: boolean}>`
+  border-radius: 50px;
+  border: 2px solid ${Colors.dark_grey};
   font-family: "Poppins", "sans-serif";
-  color: ${Colors.white};
-  background-color: ${Colors.red};
-  width: inherit;
-  border: 0;
-  border-radius: 20px;
-  font-weight: 700;
+  font-weight: 600;
   font-size: 1em;
-  padding: 8px 0px;
-  display: grid;
-  place-items: center;
-  cursor: pointer;
+  padding: 0px 40px;
+  color: ${Colors.dark_grey};
+  background-color: ${props => props.subscribed? Colors.white_1: Colors.blue_3};
+  box-shadow: 5px 5px 0 ${Colors.blue_2}, 5px 5px 0 2px ${Colors.dark_grey};
+
+  :hover {
+    background-color: ${props => props.subscribed? Colors.blue_3 : Colors.blue_accent};
+    position: relative; 
+    top: 3px;
+    left: 3px;
+    box-shadow: 2px 2px 0 ${Colors.blue_2}, 2px 2px 0 2px ${Colors.dark_grey};
+  }
 `;
 
 const ForumBackground = styled.div`
@@ -35,11 +43,13 @@ const ForumBackground = styled.div`
   width: 17.5vw;
   max-width: 17.5vw;
   padding: 1.5em;
-  border: none;
+  border: 2px solid ${Colors.dark_grey};
   border-radius: 20px;
   background-color: ${Colors.green_2};
   color: ${Colors.black};
   font-family: "Poppins", "sans-serif";
+  box-shadow: 7px 7px 0 ${Colors.blue_3},
+            7px 7px 0 2px ${Colors.dark_grey};
 `;
 
 const Top = styled.div`
@@ -126,25 +136,29 @@ const ModuleForum = ({ selectedModule }: { selectedModule: string }) => {
   };
 
   const handleButtonClick = () => {
-    setIsLoading(true)
-    fetch(API_URL + `/subscribes/${selectedModule.toUpperCase()}/${userId}`, {
-      method: isSubscribed ? "DELETE" : "POST",
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    })
-      .then((response) => response.json())
-      .then((data) => {
-        fetchMod();
-        if (isSubscribed) {
-          dispatch(deleteSubscription(selectedModule));
-        } else {
-          dispatch(addSubscription(selectedModule));
-        }
-        setIsSubscribed(!isSubscribed);
+    if (!isLoggedIn(token, userId)) {
+      dispatch(toggleLogin(true));
+    } else {
+      setIsLoading(true)
+      fetch(API_URL + `/subscribes/${selectedModule.toUpperCase()}/${userId}`, {
+        method: isSubscribed ? "DELETE" : "POST",
+        headers: {
+          Authorization: `Bearer ${token}`,
+        },
       })
-      .catch((error) => console.log(error))
-      .finally(() => setIsLoading(false));
+        .then((response) => response.json())
+        .then((data) => {
+          fetchMod();
+          if (isSubscribed) {
+            dispatch(deleteSubscription(selectedModule));
+          } else {
+            dispatch(addSubscription(selectedModule));
+          }
+          setIsSubscribed(!isSubscribed);
+        })
+        .catch((error) => console.log(error))
+        .finally(() => setIsLoading(false));
+    }
   };
 
   // useEffect(() => {
@@ -170,18 +184,17 @@ const ModuleForum = ({ selectedModule }: { selectedModule: string }) => {
           <SubscriberDiv>
             <PeopleAltIcon />
             <SubscriberDesc href={`/subscribers/${selectedModule}`}>
-              {module.SubscriberCount} subscribers
+              {module.SubscriberCount === 1? "1 subscriber" : `${module.SubscriberCount} subscribers`}
             </SubscriberDesc>
           </SubscriberDiv>
-          <RedButton onClick={handleButtonClick} disabled={isLoading}>
-            {
-              isLoading
+          <Button subscribed={isSubscribed} onClick={handleButtonClick} disabled={isLoading}>
+            { isLoading
               ? <WhiteLoader />
               : isSubscribed 
               ? <p>Unsubscribe</p> 
               : <p>Subscribe</p>
             }
-          </RedButton>
+          </Button>
         </Bottom>
       </ForumBackground>
     </ModuleForumDiv>
